Add static constraint to  overcome complex to join and retrive data. 23MAY2018

+++ src/main/java/com/programmer/gate/repository/St001Repository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.programmer.gate.model.St001;
 
 @Repository
-public interface St001Repository extends CrudRepository<St001, Double> {
+public interface St001Repository extends CrudRepository<St001, Integer> {
 
 }
